requirements.txt
django==4.2.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-environ==0.11.2
django-dotenv==1.4.2
django-phonenumber-field==7.1.0
django_phonenumbers==1.0.1
environ==1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
Pillow==10.0.1
celery==5.2.7
